requirements.txt
Django==1.10.5
djangorestframework==3.5.4
django-webpack-loader==0.4.1
django_extensions==1.7.7
honcho==0.7.1
